backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
modelz-llm==23.7.4
python-dotenv==1.0.0
redis==5.0.1
langchain==0.0.350
openai==1.3.5 